corrections

+++ qutip/core/qobj.py
@@ -828,7 +828,7 @@ class Qobj:
                         dims=self._dims,
                         isherm=self._isherm,
                         copy=False)
-        return Qobj(_data.expm(self._data, dtype=dtype),
+        return Qobj(_data.expm(self._data, dtype=self.dtype),
                     dims=self._dims,
                     isherm=self._isherm,
                     copy=False)
